Changed registration ratio to float

+++ xxclient/src/main/java/io/elixxir/xxclient/models/NodeRegistrationReport.kt
@@ -4,6 +4,6 @@ data class NodeRegistrationReport(
     val registeredCount: Long,
     val totalCount: Long
 ) : BindingsModel {
-    val ratio: Long get() = registeredCount/totalCount
+    val ratio: Float get() = registeredCount.toFloat() / totalCount.toFloat()
 }
